+++ catena/core/nodes/math/max.py
@@ -39,4 +39,4 @@ class MaxNode(CatenaNode):
             image_b = cv2.resize(image_b, (width, height))
 
         result = numpy.maximum(image_a, image_b)
-        return result
+        return result.astype(numpy.float32)
